Move benchmarks under napari package as a hidden package
