+++ napari/_benchmarks/benchmark_image_layer.py
@@ -1,28 +1,25 @@
 # See "Writing benchmarks" in the asv docs for more information.
 # https://asv.readthedocs.io/en/latest/writing_benchmarks.html
 # or the napari documentation on benchmarking
-# https://github.com/napari/napari/blob/master/BENCHMARKS.md
+# https://github.com/napari/napari/blob/master/docs/BENCHMARKS.md
 import numpy as np
-from napari.layers import Shapes
+from napari.layers import Image
 
 
-class Shapes2DSuite:
-    """Benchmarks for the Shapes layer with 2D data"""
+class Image2DSuite:
+    """Benchmarks for the Image layer with 2D data."""
 
-    params = [2 ** i for i in range(4, 9)]
+    params = [2 ** i for i in range(4, 13)]
 
     def setup(self, n):
         np.random.seed(0)
-        self.data = [50 * np.random.random((6, 2)) for i in range(n)]
-        self.layer = Shapes(self.data, shape_type='polygon')
+        self.data = np.random.random((n, n))
+        self.new_data = np.random.random((n, n))
+        self.layer = Image(self.data)
 
     def time_create_layer(self, n):
         """Time to create an image layer."""
-        Shapes(self.data, shape_type='polygon')
-
-    def time_refresh(self, n):
-        """Time to refresh view."""
-        self.layer.refresh()
+        Image(self.data)
 
     def time_set_view_slice(self, n):
         """Time to set view slice."""
@@ -35,6 +32,14 @@ class Shapes2DSuite:
     def time_get_value(self, n):
         """Time to get current value."""
         self.layer.get_value()
+
+    def time_set_data(self, n):
+        """Time to get current value."""
+        self.layer.data = self.new_data
+
+    def time_refresh(self, n):
+        """Time to refresh view."""
+        self.layer.refresh()
 
     def mem_layer(self, n):
         """Memory used by layer."""
@@ -45,23 +50,20 @@ class Shapes2DSuite:
         return self.data
 
 
-class Shapes3DSuite:
-    """Benchmarks for the Shapes layer with 3D data."""
+class Image3DSuite:
+    """Benchmarks for the Image layer with 3D data."""
 
-    params = [2 ** i for i in range(4, 9)]
+    params = [2 ** i for i in range(4, 11)]
 
     def setup(self, n):
         np.random.seed(0)
-        self.data = [50 * np.random.random((6, 3)) for i in range(n)]
-        self.layer = Shapes(self.data, shape_type='polygon')
+        self.data = np.random.random((n, n, n))
+        self.new_data = np.random.random((n, n, n))
+        self.layer = Image(self.data)
 
     def time_create_layer(self, n):
-        """Time to create a layer."""
-        Shapes(self.data, shape_type='polygon')
-
-    def time_refresh(self, n):
-        """Time to refresh view."""
-        self.layer.refresh()
+        """Time to create an image layer."""
+        Image(self.data)
 
     def time_set_view_slice(self, n):
         """Time to set view slice."""
@@ -75,9 +77,17 @@ class Shapes3DSuite:
         """Time to get current value."""
         self.layer.get_value()
 
+    def time_set_data(self, n):
+        """Time to get current value."""
+        self.layer.data = self.new_data
+
+    def time_refresh(self, n):
+        """Time to refresh view."""
+        self.layer.refresh()
+
     def mem_layer(self, n):
         """Memory used by layer."""
-        return self.layer
+        return Image(self.data)
 
     def mem_data(self, n):
         """Memory used by raw data."""
