+++ napari/_benchmarks/benchmark_surface_layer.py
@@ -1,25 +1,32 @@
 # See "Writing benchmarks" in the asv docs for more information.
 # https://asv.readthedocs.io/en/latest/writing_benchmarks.html
 # or the napari documentation on benchmarking
-# https://github.com/napari/napari/blob/master/BENCHMARKS.md
+# https://github.com/napari/napari/blob/master/docs/BENCHMARKS.md
 import numpy as np
-from napari.layers import Image
+from napari.layers import Surface
 
 
-class Image2DSuite:
-    """Benchmarks for the Image layer with 2D data."""
+class Surface2DSuite:
+    """Benchmarks for the Surface layer with 2D data"""
 
-    params = [2 ** i for i in range(4, 13)]
+    params = [2 ** i for i in range(4, 18, 2)]
 
     def setup(self, n):
         np.random.seed(0)
-        self.data = np.random.random((n, n))
-        self.new_data = np.random.random((n, n))
-        self.layer = Image(self.data)
+        self.data = (
+            np.random.random((n, 2)),
+            np.random.randint(n, size=(n, 3)),
+            np.random.random(n),
+        )
+        self.layer = Surface(self.data)
 
     def time_create_layer(self, n):
         """Time to create an image layer."""
-        Image(self.data)
+        Surface(self.data)
+
+    def time_refresh(self, n):
+        """Time to refresh view."""
+        self.layer.refresh()
 
     def time_set_view_slice(self, n):
         """Time to set view slice."""
@@ -32,14 +39,6 @@ class Image2DSuite:
     def time_get_value(self, n):
         """Time to get current value."""
         self.layer.get_value()
-
-    def time_set_data(self, n):
-        """Time to get current value."""
-        self.layer.data = self.new_data
-
-    def time_refresh(self, n):
-        """Time to refresh view."""
-        self.layer.refresh()
 
     def mem_layer(self, n):
         """Memory used by layer."""
@@ -50,20 +49,27 @@ class Image2DSuite:
         return self.data
 
 
-class Image3DSuite:
-    """Benchmarks for the Image layer with 3D data."""
+class Surface3DSuite:
+    """Benchmarks for the Surface layer with 3D data."""
 
-    params = [2 ** i for i in range(4, 11)]
+    params = [2 ** i for i in range(4, 18, 2)]
 
     def setup(self, n):
         np.random.seed(0)
-        self.data = np.random.random((n, n, n))
-        self.new_data = np.random.random((n, n, n))
-        self.layer = Image(self.data)
+        self.data = (
+            np.random.random((n, 3)),
+            np.random.randint(n, size=(n, 3)),
+            np.random.random(n),
+        )
+        self.layer = Surface(self.data)
 
     def time_create_layer(self, n):
-        """Time to create an image layer."""
-        Image(self.data)
+        """Time to create a layer."""
+        Surface(self.data)
+
+    def time_refresh(self, n):
+        """Time to refresh view."""
+        self.layer.refresh()
 
     def time_set_view_slice(self, n):
         """Time to set view slice."""
@@ -77,17 +83,9 @@ class Image3DSuite:
         """Time to get current value."""
         self.layer.get_value()
 
-    def time_set_data(self, n):
-        """Time to get current value."""
-        self.layer.data = self.new_data
-
-    def time_refresh(self, n):
-        """Time to refresh view."""
-        self.layer.refresh()
-
     def mem_layer(self, n):
         """Memory used by layer."""
-        return Image(self.data)
+        return self.layer
 
     def mem_data(self, n):
         """Memory used by raw data."""
