+++ napari/_benchmarks/benchmark_vectors_layer.py
@@ -1,24 +1,24 @@
 # See "Writing benchmarks" in the asv docs for more information.
 # https://asv.readthedocs.io/en/latest/writing_benchmarks.html
 # or the napari documentation on benchmarking
-# https://github.com/napari/napari/blob/master/BENCHMARKS.md
+# https://github.com/napari/napari/blob/master/docs/BENCHMARKS.md
 import numpy as np
-from napari.layers import Points
+from napari.layers import Vectors
 
 
-class Points2DSuite:
-    """Benchmarks for the Points layer with 2D data"""
+class Vectors2DSuite:
+    """Benchmarks for the Vectors layer with 2D data"""
 
     params = [2 ** i for i in range(4, 18, 2)]
 
     def setup(self, n):
         np.random.seed(0)
-        self.data = np.random.random((n, 2))
-        self.layer = Points(self.data)
+        self.data = np.random.random((n, 2, 2))
+        self.layer = Vectors(self.data)
 
     def time_create_layer(self, n):
-        """Time to create layer."""
-        Points(self.data)
+        """Time to create an image layer."""
+        Vectors(self.data)
 
     def time_refresh(self, n):
         """Time to refresh view."""
@@ -35,6 +35,14 @@ class Points2DSuite:
     def time_get_value(self, n):
         """Time to get current value."""
         self.layer.get_value()
+
+    def time_width(self, n):
+        """Time to update width."""
+        self.layer.width = 2
+
+    def time_length(self, n):
+        """Time to update length."""
+        self.layer.length = 2
 
     def mem_layer(self, n):
         """Memory used by layer."""
@@ -45,19 +53,19 @@ class Points2DSuite:
         return self.data
 
 
-class Points3DSuite:
-    """Benchmarks for the Points layer with 3D data."""
+class Vectors3DSuite:
+    """Benchmarks for the Vectors layer with 3D data."""
 
     params = [2 ** i for i in range(4, 18, 2)]
 
     def setup(self, n):
         np.random.seed(0)
-        self.data = np.random.random((n, 3))
-        self.layer = Points(self.data)
+        self.data = np.random.random((n, 2, 3))
+        self.layer = Vectors(self.data)
 
     def time_create_layer(self, n):
-        """Time to create layer."""
-        Points(self.data)
+        """Time to create a layer."""
+        Vectors(self.data)
 
     def time_refresh(self, n):
         """Time to refresh view."""
@@ -74,6 +82,14 @@ class Points3DSuite:
     def time_get_value(self, n):
         """Time to get current value."""
         self.layer.get_value()
+
+    def time_width(self, n):
+        """Time to update width."""
+        self.layer.width = 2
+
+    def time_length(self, n):
+        """Time to update length."""
+        self.layer.length = 2
 
     def mem_layer(self, n):
         """Memory used by layer."""
